+++ verik-compiler/regression/08-class.sv
@@ -23,15 +23,8 @@
 
 class C0;
 
-    static function automatic C0 __new();
-        C0 __0;
-        __0 = new();
-        __0.__init();
-        return __0;
-    endfunction : __new
-
-    function automatic void __init();
-    endfunction : __init
+    function new();
+    endfunction : new
 
     static logic x4 = 1'b0;
 
@@ -39,16 +32,9 @@ endclass : C0
 
 class C1 extends C0;
 
-    static function automatic C1 __new();
-        C1 __1;
-        __1 = new();
-        __1.__init();
-        return __1;
-    endfunction : __new
-
-    function automatic void __init();
-        super.__init();
-    endfunction : __init
+    function new();
+        super.new();
+    endfunction : new
 
 endclass : C1
 
@@ -60,23 +46,16 @@ endclass : C2
 
 class C2_T_Int;
 
-    static function automatic C2_T_Int __new();
-        C2_T_Int __2;
-        __2 = new();
-        __2.__init();
-        return __2;
-    endfunction : __new
-
-    function automatic void __init();
-    endfunction : __init
+    function new();
+    endfunction : new
 
 endclass : C2_T_Int
 
 module M;
 
-    C0       x0 = C0::__new();
-    C1       x1 = C1::__new();
-    C2_T_Int x2 = C2_T_Int::__new();
+    C0       x0 = C0::new();
+    C1       x1 = C1::new();
+    C2_T_Int x2 = C2_T_Int::new();
 
     initial begin : f0
         $display($sformatf("%b", C0::x4));