+++ verik-compiler/src/main/kotlin/io/verik/compiler/ast/element/declaration/sv/ESvConstructor.kt
@@ -20,30 +20,28 @@ import io.verik.compiler.ast.common.Type
 import io.verik.compiler.ast.element.declaration.common.ETypeParameter
 import io.verik.compiler.ast.element.expression.common.EBlockExpression
 import io.verik.compiler.ast.property.AnnotationEntry
-import io.verik.compiler.ast.property.FunctionQualifierType
 import io.verik.compiler.common.Visitor
 import io.verik.compiler.message.SourceLocation
 
-class ESvFunction(
+class ESvConstructor(
     override val location: SourceLocation,
-    override var name: String,
     override var type: Type,
     override var annotationEntries: List<AnnotationEntry>,
     override var documentationLines: List<String>?,
     override var body: EBlockExpression,
-    override var typeParameters: ArrayList<ETypeParameter>,
-    override var valueParameters: ArrayList<ESvValueParameter>,
-    override val isStatic: Boolean,
-    val qualifierType: FunctionQualifierType
+    override var valueParameters: ArrayList<ESvValueParameter>
 ) : ESvAbstractFunction() {
+
+    override var name = "new"
+    override var typeParameters: ArrayList<ETypeParameter> = ArrayList()
+    override val isStatic = true
 
     init {
         body.parent = this
-        typeParameters.forEach { it.parent = this }
         valueParameters.forEach { it.parent = this }
     }
 
     override fun accept(visitor: Visitor) {
-        return visitor.visitSvFunction(this)
+        visitor.visitSvConstructor(this)
     }
 }