+++ verik-compiler/src/main/kotlin/io/verik/compiler/common/ElementPrinter.kt
@@ -45,6 +45,7 @@ import io.verik.compiler.ast.element.declaration.sv.EModulePortInstantiation
 import io.verik.compiler.ast.element.declaration.sv.EPort
 import io.verik.compiler.ast.element.declaration.sv.EStruct
 import io.verik.compiler.ast.element.declaration.sv.ESvClass
+import io.verik.compiler.ast.element.declaration.sv.ESvConstructor
 import io.verik.compiler.ast.element.declaration.sv.ESvFunction
 import io.verik.compiler.ast.element.declaration.sv.ESvValueParameter
 import io.verik.compiler.ast.element.declaration.sv.ETask
@@ -250,27 +251,6 @@ class ElementPrinter : Visitor() {
         }
     }
 
-    override fun visitSvFunction(function: ESvFunction) {
-        build("SvFunction") {
-            build(function.name)
-            build(function.type.toString())
-            build(function.body)
-            build(function.typeParameters)
-            build(function.valueParameters)
-            build(function.qualifierType.toString())
-            build(function.isStatic)
-        }
-    }
-
-    override fun visitTask(task: ETask) {
-        build("Task") {
-            build(task.name)
-            build(task.body)
-            build(task.typeParameters)
-            build(task.valueParameters)
-        }
-    }
-
     override fun visitPrimaryConstructor(primaryConstructor: EPrimaryConstructor) {
         build("PrimaryConstructor") {
             build(primaryConstructor.name)
@@ -287,6 +267,36 @@ class ElementPrinter : Visitor() {
             build(secondaryConstructor.body)
             build(secondaryConstructor.valueParameters)
             build(secondaryConstructor.superTypeCallExpression)
+        }
+    }
+
+    override fun visitSvFunction(function: ESvFunction) {
+        build("SvFunction") {
+            build(function.name)
+            build(function.type.toString())
+            build(function.body)
+            build(function.typeParameters)
+            build(function.valueParameters)
+            build(function.isStatic)
+            build(function.qualifierType.toString())
+        }
+    }
+
+    override fun visitTask(task: ETask) {
+        build("Task") {
+            build(task.name)
+            build(task.body)
+            build(task.typeParameters)
+            build(task.valueParameters)
+            build(task.isStatic)
+        }
+    }
+
+    override fun visitSvConstructor(constructor: ESvConstructor) {
+        build("SvConstructor") {
+            build(constructor.type.toString())
+            build(constructor.body)
+            build(constructor.valueParameters)
         }
     }