+++ verik-compiler/src/main/kotlin/io/verik/compiler/common/Visitor.kt
@@ -59,6 +59,7 @@ import io.verik.compiler.ast.element.declaration.sv.EPort
 import io.verik.compiler.ast.element.declaration.sv.EStruct
 import io.verik.compiler.ast.element.declaration.sv.ESvAbstractFunction
 import io.verik.compiler.ast.element.declaration.sv.ESvClass
+import io.verik.compiler.ast.element.declaration.sv.ESvConstructor
 import io.verik.compiler.ast.element.declaration.sv.ESvFunction
 import io.verik.compiler.ast.element.declaration.sv.ESvValueParameter
 import io.verik.compiler.ast.element.declaration.sv.ETask
@@ -244,6 +245,10 @@ abstract class Visitor {
 
     open fun visitTask(task: ETask) {
         visitSvAbstractFunction(task)
+    }
+
+    open fun visitSvConstructor(constructor: ESvConstructor) {
+        visitSvAbstractFunction(constructor)
     }
 
     open fun visitAbstractProceduralBlock(abstractProceduralBlock: EAbstractProceduralBlock) {