+++ verik-compiler/src/main/kotlin/io/verik/compiler/serialize/source/DeclarationSerializer.kt
@@ -34,6 +34,7 @@ import io.verik.compiler.ast.element.declaration.sv.EPort
 import io.verik.compiler.ast.element.declaration.sv.EStruct
 import io.verik.compiler.ast.element.declaration.sv.ESvAbstractFunction
 import io.verik.compiler.ast.element.declaration.sv.ESvClass
+import io.verik.compiler.ast.element.declaration.sv.ESvConstructor
 import io.verik.compiler.ast.element.declaration.sv.ESvFunction
 import io.verik.compiler.ast.element.declaration.sv.ESvValueParameter
 import io.verik.compiler.ast.element.declaration.sv.ETask
@@ -163,6 +164,17 @@ object DeclarationSerializer {
         }
         serializeContext.label(task.body.endLocation) {
             serializeContext.appendLine("endtask : ${task.name}")
+        }
+    }
+
+    fun serializeConstructor(constructor: ESvConstructor, serializeContext: SerializeContext) {
+        serializeContext.append("function new")
+        serializeValueParameterList(constructor, serializeContext)
+        serializeContext.indent {
+            constructor.body.statements.forEach { serializeContext.serializeAsStatement(it) }
+        }
+        serializeContext.label(constructor.body.endLocation) {
+            serializeContext.appendLine("endfunction : new")
         }
     }